+++ Roguelike/src/Object/Tiles/Tile.h
@@ -5,9 +5,9 @@ namespace CR::Objects {
 	class Tile : public GameObject {
 	public:
 		Tile(char skin, unsigned short skinColor, int x, int y)
-			: GameObject(skin, skinColor, x, y) {}
+			: GameObject(skin, skinColor, (float) x, (float) y) {}
 		
 		Tile(char skin, unsigned short skinColor, const Vector2<int>& pos)
-			: GameObject(skin, skinColor, pos.x, pos.y) {}
+			: GameObject(skin, skinColor, (float) pos.x, (float) pos.y) {}
 	};
 }
